skills/runtime/dreamhost-manager/requirements.txt
paramiko>=3.0.0
python-dotenv>=1.0.0
scp>=0.14.5
